+++ docs/source/whats-new.rst
@@ -39,6 +39,9 @@ Breaking changes
 Deprecations
 ~~~~~~~~~~~~
 
+- Renamed `DataTree.lineage` to `DataTree.parents` to match `pathlib` vocabulary
+  (:issue:`283`, :pull:`286`). `lineage` is now deprecated and use of `parents` is encouraged.
+
 Bug fixes
 ~~~~~~~~~
 - Keep attributes on nodes containing no data in :py:func:`map_over_subtree`. (:issue:`278`, :pull:`279`)